feat: reestucture dinamic dashboard
