--- src/app/pages/admin/dashboard/dashboard.component.html
@@ -1,18 +1,0 @@
-<div>
-    <div class="title">
-        <img src="../../../assets/imagenes/logo.png" alt="">
-        <div class="welcome">
-            <div>
-                <h1>BIENVENIDO!!!</h1>
-            </div>
-            <div class="role">
-                <h3>Administrador</h3>
-            </div>
-        </div>
-    </div>
-    <br>
-    <div class="links">
-        <a [routerLink]="'/selected-role'">Agregar empleado</a>
-        <a [routerLink]="">Lista de empleados</a>
-    </div>
-</div>
