--- src/app/pages/administrative/administrative.component.html
@@ -1,18 +1,0 @@
-<div>
-    <div class="title">
-        <img src="../../../assets/imagenes/logo.png" alt="">
-        <div class="welcome">
-            <div>
-                <h1>BIENVENIDO!!!</h1>
-            </div>
-            <div class="role">
-                <h3>Administrativo</h3>
-            </div>
-        </div>
-    </div>
-    <br>
-    <div class="links">
-        <a href="">Agregar empleado</a>
-        <a href="">Lista de empleados</a>
-    </div>
-</div>
