--- src/app/pages/administrative/administrative.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-administrative',
-  templateUrl: './administrative.component.html',
-  styleUrl: './administrative.component.css'
-})
-export class AdministrativeComponent {
-
-}
