+++ src/app/pages/dashboard/dashboard.component.ts
@@ -2,30 +2,28 @@ import { Component, OnInit } from '@angular/core';
 import { LoginService } from '../../services/login.service';
 
 @Component({
-  selector: 'app-navbar',
-  templateUrl: './navbar.component.html',
-  styleUrl: './navbar.component.css'
+  selector: 'app-dashboard',
+  templateUrl: './dashboard.component.html',
+  styleUrl: './dashboard.component.css'
 })
-export class NavbarComponent implements OnInit{
+export class DashboardComponent implements OnInit {
+  isLoggedIn: boolean = false;
+  user: any = null;
+  role: any = null;
 
-  isLoggedIn = false;
-  user:any = null;
   constructor(public login: LoginService){}
 
   ngOnInit(): void {
     this.isLoggedIn = this.login.isLoggedIn();
     this.user = this.login.getUser();
+    this.role = this.login.getUserRole();
     this.login.loginStatusSubjec.asObservable().subscribe(
       data => {
         this.isLoggedIn = this.login.isLoggedIn();
         this.user = this.login.getUser();
+        this.role = this.login.getUserRole();
       }
     )
   }
-
-  public logout(){
-    this.login.logout();
-    window.location.reload();
-  }
-
 }
+
