--- src/app/pages/user/user-dashboard/user-dashboard.component.css
@@ -1,31 +1,0 @@
-.title {
-    display: flex !important;
-}
-.title img {
-    padding: 1.5rem !important;
-}
-.welcome {
-    padding: 0.5rem !important;
-    display: flex !important;
-    flex-direction: column !important;
-    justify-content: center !important;
-    align-items: center !important;
-}
-.welcome h1{
-    font-size: 70px !important;
-}
-.links {
-    display: inline-flex !important;
-    flex-direction: column !important;
-    padding: 0.5rem 1rem !important;
-}
-.links a {
-    margin: 0.5rem !important;
-    font-size: 20px !important;
-}
-.role {
-    display: flex !important;
-    flex-direction: column !important;
-    padding: 0.5rem 1rem !important;
-    width: 100%;
-}
