--- src/app/pages/user/user-dashboard/user-dashboard.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-user-dashboard',
-  templateUrl: './user-dashboard.component.html',
-  styleUrl: './user-dashboard.component.css'
-})
-export class UserDashboardComponent {
-
-}
