+++ src/app/shared/navbar/navbar.component.html
@@ -1,0 +1,24 @@
+<mat-toolbar color="primary">
+    <a [routerLink]="'/'" mat-icon-button class="example-icon" aria-label="Example icon-button with menu icon">
+      <mat-icon>home</mat-icon>
+    </a>
+    <span>Optica Yara</span>
+    <span class="example-spacer"></span>
+    <a *ngIf="!isLoggedIn" [routerLink]="'/login'" mat-icon-button class="example-icon favorite-icon"
+      aria-label="Example icon-button with heart icon">
+      Iniciar sesión
+    </a>
+    <a *ngIf="!isLoggedIn" [routerLink]="'/signup'" mat-icon-button class="example-icon favorite-icon mr20 ml70"
+      aria-label="Example icon-button with heart icon">
+      Registrar
+    </a>
+    <a *ngIf="isLoggedIn" (click)="logout()" style="margin-right: 50px" mat-icon-button class="example-icon favorite-icon"
+      aria-label="Example icon-button with heart icon">
+      Cerrar sesión
+    </a>
+    <button *ngIf="isLoggedIn && user" mat-button style="margin-right: 10px" class="example-icon favorite-icon"
+      aria-label="Example icon-button with heart icon" [routerLink]="'/admin/profile'">
+      {{ user.username }}
+    </button>
+  </mat-toolbar>
+
