+++ src/app/shared/navbar/navbar.component.ts
@@ -1,15 +1,15 @@
 import { Component, OnInit } from '@angular/core';
-import { LoginService } from '../../../services/login.service';
+import { LoginService } from '../../services/login.service';
 
 @Component({
-  selector: 'app-dashboard',
-  templateUrl: './dashboard.component.html',
-  styleUrl: './dashboard.component.css'
+  selector: 'app-navbar',
+  templateUrl: './navbar.component.html',
+  styleUrl: './navbar.component.css'
 })
-export class DashboardComponent implements OnInit {
-  isLoggedIn: boolean = false;
-  user: any = null;
+export class NavbarComponent implements OnInit{
 
+  isLoggedIn = false;
+  user:any = null;
   constructor(public login: LoginService){}
 
   ngOnInit(): void {
@@ -22,4 +22,10 @@ export class DashboardComponent implements OnInit {
       }
     )
   }
+
+  public logout(){
+    this.login.logout();
+    window.location.reload();
+  }
+
 }
